api/requirements.txt
requests
certifi
bs4
flask
pandas
numpy
scikit-learn
nltk
joblib
gunicorn
mysql.connector
flask_cors